Refactor some of the sim types: coordinates, world shape
Use a proper Point type for coordinates, and a common world shape type
for the size of the simulated world.

Switch the P2P sim code from using DiffTime to Double as the
representation of Latency. Also switch from using a (1,1) unit square as
the virtual simulated world dimensions, to using the dimensions given as
the latency and returning that in the P2PTopography. This eliminates a
whole set of coordinate transformations, which is a source of bugs.