WIP frisby-sampler.html
A static Dynamic HTML file that generates and renders a Leios chain schema.

It's just for pondering thinking and was a personal exercise to implement. For
example, implementing it helped me realize that it'd be plausible for an EB to
include an older EB that it hasn't seen the cert for if it had seen the cert
for another in-between EB that also includes the older EB.