Create a dedicated threat model for Leios (#452)
* Create a threat model for Leios

This is a first sketch for a threat model as a dedicated artifact. We
have another similar collection of threats here

https://github.com/input-output-hk/ouroboros-leios/blob/main/docs/technical-report-1.md#threat-model

In this document, we follow an established process of threat modeling by
also describing a system overview, assets to protect and go more into
detail detail on threats, prerequisites, attack vector, cost
and impact.

This should be updated with more detail / reference the concrete
protocol variant we propose in the CIP and accordingly with assets,
threats and mitigations.

* Fix terminology and be less concrete on committe and quorum size

* Add links to related documents and fix formatting

* Drop headline numbering

* Fix headings, enumerate assets and rename M5

* Updates to T1, M1 and M2

* Differentiate between downstream and upstream

* Sketch two more threats and update document status

* Add the honey pot threat

* Add a warning to risk assesment matrix