Update website: next monthly and deduplicate pages (#540)
* Allow for exceptions in next monthly meeting calculation

The monthly review meeting for september will not be on the last
wednesday of the month, but shifted by one week to Oct 1st.

* Drop protocol overview and explainers in favor of landing page

The landing page now gives a better explanation of how leios works and
points to the CIP which is the most up-to-date resource we have.

* Update resources

Several links were broken, outdated or already included in other parts
of the website.

* Update logbook